+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,9 +2,9 @@
 using System;
 public interface IDamageable
 {
-    public int Lives { get; }
-    public int HP { get; }
+    public sbyte Lives { get; }
+    public sbyte HP { get; }
     public bool IsPlayer { get; }
 
-    public void Damage(int hpDamage);
+    public void Damage(sbyte hpDamage);
 }
