+++ Assets/Scripts/Managers/MainMenu.cs
@@ -6,7 +6,7 @@ namespace Veganimus.Platformer
         [SerializeField] private SceneLoader _sceneLoader;
         [SerializeField] private ExitButton exitButton;
 
-        public void LoadScene(int sceneIndex)
+        public void LoadScene(byte sceneIndex)
         {
             _sceneLoader.LoadGame(sceneIndex);
         }
